move the dictionary context out of the header
